Edited the Install file. It now works on my computer at the very least.
As mentioned in the title I tweaked the installation file.

There were several problems I found and fixed.

Without the "" around the directories, any space would not be included. Therefore a path such as C:\Users\Nils Schofer\ will cause the install to break. The "" seemed to have fixed this problem.

The set modpackFolder=... also didn't seem be being called. When echo %modpackFolder% it returned nothing. I moved that definition to the begining and it seemed to have fixed the problem.

Finally, changing every ! to % which fixed some issues as well.

Some things to mention:
- You have to run Forge at least once for it to add the jar file to the versions folder. Therefore, the code will ask you to reinstall forge every time you run it until you have opened minecraft and run Forge at least once.
- I can't seem to run the game. I added the profile (copied what you left commented out) to my minecraft launcher. I also realized that Minecraft Launcher already adds %Appdata%\.minecraft when following  the "gameDir". I changed the commented out version of the copy to match this: "gameDir" : "profiles\\"GAMES_Modpack1.0". Unfortunately, as previously stated the game still does not run. It simply gives me a Error Code 0.