Created a .bat file with code to input Profile
As mentioned in the file, it copies everything to a temporary file then renames the new file. NOTE: It can't copy the Icon for Forge (for some reason).

We can either have the install.bat run this code, or simply add the code to the Install.bat file.

Let me know if it works : D